Implementaion
The Mandelbrot set is one of the most complex and beautiful images arising from mathematics. It is defined to be the set of complex numbers for which the function fc(z) = z2 + c does not diverge when iterated from z=0. For a clear explanation of what that means, see this video on YouTube featuring Dr. Holly Krieger from MIT.